+++ src/main/resources/assets/trafficcontrol/blockstates/ped_crossing_lamps_sw_lamp.json
@@ -1,0 +1,43 @@
+{
+    "forge_marker": 1,
+    "defaults": {
+        "textures": {
+            "leftlamp": "trafficcontrol:blocks/lamp_off",
+            "rightlamp": "trafficcontrol:blocks/lamp_off"
+        }
+    },
+    "variants": {
+        "rotation": {
+            "0": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp" },
+            "1": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 337 } } },
+            "2": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 315 } } },
+            "3": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 292 } } },
+            "4": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 270 } } },
+            "5": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 247 } } },
+            "6": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 225 } } },
+            "7": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 202 } } },
+            "8": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 180 } } },
+            "9": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 157 } } },
+            "10": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 135 } } },
+            "11": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 112 } } },
+            "12": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 90 } } },
+            "13": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 67 } } },
+            "14": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 45 } } },
+            "15": { "model": "trafficcontrol:ped_crossing_lamps_sw_lamp", "transform": { "rotation": { "y": 22 } } },
+            "16": { }
+        },
+        "state": {
+            "off": {},
+            "flash1": {
+                "textures": {
+                    "leftlamp": "trafficcontrol:blocks/red"
+                }
+            },
+            "flash2": {
+                "textures": {
+                    "rightlamp": "trafficcontrol:blocks/red"
+                }
+            }
+        }
+    }
+}
