+++ src/main/resources/assets/trafficcontrol/models/block/ped_crossing_lamps_nw_lamp.json
@@ -1,0 +1,22 @@
+{
+	"parent": "block/block",
+	"textures": {
+		"particle": "trafficcontrol:blocks/generic",
+		"leftlamptex": "#leftlamp"
+	},
+	"elements": [
+		{
+			"name": "Lens NW",
+			"from": [2, 10, 6.98],
+			"to": [5, 13, 7.02],
+			"faces": {
+				"north": {"uv": [0, 0, 3, 3], "texture": "#leftlamptex"},
+				"east": {"uv": [0, 0, 1, 3], "texture": "#leftlamptex"},
+				"south": {"uv": [0, 0, 3, 3], "texture": "#leftlamptex"},
+				"west": {"uv": [0, 0, 1, 3], "texture": "#leftlamptex"},
+				"up": {"uv": [0, 0, 3, 1], "texture": "#leftlamptex"},
+				"down": {"uv": [0, 0, 3, 1], "texture": "#leftlamptex"}
+			}
+		}
+	]
+}
