+++ src/main/resources/assets/trafficcontrol/models/block/ped_crossing_lamps_sw_lamp.json
@@ -1,0 +1,22 @@
+{
+	"parent": "block/block",
+	"textures": {
+		"particle": "trafficcontrol:blocks/generic",
+		"rightlamptex": "#rightlamp"
+	},
+	"elements": [
+		{
+			"name": "Lens SW",
+			"from": [11, 10, 6.98],
+			"to": [14, 13, 7.02],
+			"faces": {
+				"north": {"uv": [0, 0, 3, 3], "texture": "#rightlamptex"},
+				"east": {"uv": [0, 0, 1, 3], "texture": "#rightlamptex"},
+				"south": {"uv": [0, 0, 3, 3], "texture": "#rightlamptex"},
+				"west": {"uv": [0, 0, 1, 3], "texture": "#rightlamptex"},
+				"up": {"uv": [0, 0, 3, 1], "texture": "#rightlamptex"},
+				"down": {"uv": [0, 0, 3, 1], "texture": "#rightlamptex"}
+			}
+		}
+	]
+}
